donnie_bot/requirements.txt
# Core Discord
discord.py>=2.3.0

# Database
aiosqlite>=0.19.0

# AI Service
anthropic>=0.25.0

# Voice/Audio (optional, but needed for framework)
PyNaCl>=1.5.0

# Configuration
python-dotenv>=1.0.0

# Caching
cachetools>=5.3.0

# Utilities
aiofiles>=23.0.0